requirements.txt
Django==5.0.1
django-environ==0.11.2
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
Pillow
requests==2.31.0